register-login buttons enable works

+++ backend/api/users/user-exists.php
@@ -3,22 +3,26 @@ require_once("../../db/db.php");
 //    session_start();
         try{
             $userData = json_decode(file_get_contents("php://input"), true);
-            var_dump($userData);
             $db = new DB();
             $connection = $db->getConnection();
             $sql = "SELECT username, email FROM users WHERE username=:username OR email=:email";
             $statement = $connection -> prepare($sql);
             $statement -> execute(["username" => $userData["username"], "email" => $userData["email"]]);
             $results = $statement->fetchAll(PDO::FETCH_ASSOC);
+            $exists = false;
             while ($user = $results) {
                 if($user["email"] == $userData["email"]) {
+                    $exists = true;
                     echo json_encode(array("exists" => true, "sameUsername"=> false, "sameEmail"=> true));
                 }
-                if($user["username"] == $userData["username"]) {
+                else if($user["username"] == $userData["username"]) {
+                    $exists = true;
                     echo json_encode(array("exists" => true, "sameUsername"=> true, "sameEmail"=> false));
                 }
             }
-            echo json_encode(array("exists" => false, "sameUsername"=> false, "sameEmail"=> false));
+            if(!$exists){
+                echo json_encode(array("exists" => false, "sameUsername"=> false, "sameEmail"=> false));
+            }
        } catch(PDOException $exc) {
            http_response_code(500);
            echo ["message" => $exc->getMessage()];
